+++ libs/sdk-ui-filters/src/AttributeFilter/AttributeFilterButton/constants.ts
@@ -1,0 +1,11 @@
+// (C) 2022 GoodData Corporation
+
+import { MAX_SELECTION_SIZE } from "../AttributeDropdown/AttributeDropdownList";
+
+export const ATTRIBUTE_FILTER_BITTON_TOOLTIP_ALIGN_POINT = [
+    { align: "tc bc", offset: { x: 0, y: -2 } },
+    { align: "cc tc", offset: { x: 0, y: 10 } },
+    { align: "bl tr", offset: { x: -2, y: -8 } },
+];
+
+export const ATTRIBUTE_FILTER_BUTTON_LIMIT = MAX_SELECTION_SIZE + 50;
